[advanced-reboot] move Arista class to seperate module
